review-like-api: fix namespace
