+++ routes/api.php
@@ -34,8 +34,8 @@ Route::prefix('v1')->group(function () {
 
             Route::post('/places/{id}/like', 'Api\LikeController@likePlace')->name('place.like');
             Route::post('/places/{id}/dislike', 'Api\DisikeController@dislikePlace')->name('place.dislike');
-            Route::post('/reviews/{id}/like', 'Api\Review\LikeReviewController@likeReview');
-            Route::post('/reviews/{id}/dislike', 'Api\Review\DisikeReviewController@dislikeReview');            
+            Route::post('/reviews/{id}/like', 'Api\LikeController@likeReview')->name('review.like');
+            Route::post('/reviews/{id}/dislike', 'Api\DisikeController@dislikeReview')->name('review.dislike');
         });
     });
 });
